--- frontend/TAI-POC/src/app/app.component.html
@@ -1,2 +1,0 @@
-
-<router-outlet></router-outlet>
