--- frontend/TAI-POC/src/app/cutomer-history/cutomer-history.component.css
@@ -1,5 +1,0 @@
-.container{
-    margin-top: 20px;
-    
-
-}
