--- frontend/TAI-POC/src/app/cutomer-history/cutomer-history.component.html
@@ -1,40 +1,0 @@
-<div class="container">
-    <table class="table table-striped">
-        <thead>
-          <tr>
-            <th scope="col">Sno</th>
-            <th scope="col">---</th>
-            <th scope="col">---</th>
-            <th scope="col">---</th>
-            <th scope="col">--</th>
-            <th scope="col">---</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr>
-            <th scope="row">1</th>
-            <td>9102-233</td>
-            <td>Otto</td>
-            <td>D5 Dallas</td>
-            <td>Dallas</td>
-            <td>PF</td>
-          </tr>
-          <tr>
-            <th scope="row">2</th>
-            <td>42123-223</td>
-            <td>Thornton</td>
-            <td>c11 California</td>
-            <td>San-francisco</td>
-            <td>PF</td>
-          </tr>
-          <tr>
-            <th scope="row">3</th>
-            <td>98723-12</td>
-            <td>Prince</td>
-            <td>c11 Boston</td>
-            <td>Massachussetes</td>
-            <td>PF</td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
