--- frontend/TAI-POC/src/app/cutomer-history/cutomer-history.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-cutomer-history',
-  templateUrl: './cutomer-history.component.html',
-  styleUrls: ['./cutomer-history.component.css']
-})
-export class CutomerHistoryComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
