--- frontend/TAI-POC/src/app/mainpage/mainpage.component.html
@@ -1,25 +1,0 @@
-<div class="main">
-    <div class="intro-div">
-        <span class="heading">Remember</span>
-        <span class="sub-heading">this is just a proof of concept, don't save production data</span>
-        <span>
-            <a href="https://github.com/cybergroupdevs/TAI-POC">
-                <button class="know-more">
-                    Code peek
-                </button>
-            </a>
-        </span>
-    </div>
-</div>
-<div class="container">
-    <div class="body">
-        <div class="col-md-5 col-7">
-            <h1>Enter transaction id</h1>
-            <form [formGroup]="transactionForm">
-            <input class="form-control" type="text" placeholder="Transaction Id" formControlName="transactionId"
-                style="text-align: right;" />
-            <button type="button" class="btn btn-dark" (click)="getTransactionDetails()">Submit</button>
-            </form>
-        </div>
-    </div>
-</div>
