--- frontend/TAI-POC/src/app/transaction-detail/consultantDialog.html
@@ -1,17 +1,0 @@
-<div class="table">
-    <table class="table table-striped">
-        <thead>
-          <tr>
-            <th scope="col">Sno</th>
-            <th scope="col">Operator</th>
-            <th scope="col">Salescns</th>
-          </tr>
-        </thead>
-        <tbody>
-            <tr>
-                <th scope="row">1</th>
-                <td>Drip</td>
-                <td>9588212</td>
-        </tbody>
-        </table>
-    </div>
