--- frontend/TAI-POC/src/app/transaction-detail/customerDialog.html
@@ -1,51 +1,0 @@
-<div class="table">
-<table class="table table-striped">
-    <thead>
-      <tr>
-        <th scope="col">Sno</th>
-        <th scope="col">Phoneno</th>
-        <th scope="col">Name</th>
-        <th scope="col">Address</th>
-        <th scope="col">City</th>
-        <th scope="col">Zip</th>
-        <th scope="col">Alt phoneno</th>
-        <th scope="col">PFID</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <th scope="row">1</th>
-        <td>9102-233</td>
-        <td>Otto</td>
-        <td>D5 Dallas</td>
-        <td>Dallas</td>
-        <td>1233341</td>
-        <td>12323-13</td>
-        <td>24555212</td>
-      </tr>
-      <tr>
-        <th scope="row">2</th>
-        <td>42123-223</td>
-        <td>Thornton</td>
-        <td>c11 California</td>
-        <td>San-francisco</td>
-        <td>7800098</td>
-        <td>1288770-12</td>
-        <td>67888807</td>
-      </tr>
-      <tr>
-        <th scope="row">3</th>
-        <td>98723-12</td>
-        <td>Prince</td>
-        <td>c11 Boston</td>
-        <td>Massachussetes</td>
-        <td>7800098</td>
-        <td>1288770-12</td>
-        <td>67888807</td>
-      </tr>
-      <tr>
-        <a [routerLink]="['customer/history']"><button type="button" class="btn btn-secondary" (click)="onNoClick()">History</button></a>
-      </tr>
-    </tbody>
-  </table>
-</div>
