--- frontend/TAI-POC/src/app/transaction-detail/transaction-detail.component.css
@@ -1,19 +1,0 @@
-table{
-    width: 100%;
-}
-
-.container{
-    width: 1000px!important;
-}
-
-.container-fluid{
-    padding: 0!important;
-}
-
-.wrapper{
-    margin-top: 10px;
-    margin-bottom: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
