+++ frontend/src/app/app.module.ts
@@ -4,7 +4,7 @@ import {MatTableModule} from '@angular/material/table';
 import {MatDialogModule} from '@angular/material/dialog';
 import { HttpClientModule } from '@angular/common/http';
 import {MatTabsModule} from '@angular/material/tabs';
-
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { TransactionDetailComponent} from './transaction-detail/transaction-detail.component';
@@ -15,6 +15,7 @@ import { CustomerComponent } from './customer/customer.component';
 import { ConsultantComponent } from './consultant/consultant.component';
 import { CustomerHistoryComponent } from './customer-history/customer-history.component';
 import { CreditCardComponent } from './credit-card/credit-card.component';
+import {Observable} from 'rxjs';
 
 @NgModule({
   declarations: [
@@ -35,7 +36,9 @@ import { CreditCardComponent } from './credit-card/credit-card.component';
     MatTableModule,
     MatDialogModule,
     HttpClientModule,
-    MatTabsModule 
+    MatTabsModule,
+    FormsModule,
+    ReactiveFormsModule
   ],
   providers: [],
   bootstrap: [AppComponent],
