+++ frontend/src/app/transaction-detail/transaction-detail.component.html
@@ -14,7 +14,7 @@
               <p>Transaction</p>
             </div>
             <div class="col-md-6">
-              <p># 3304*32*6579</p>
+              <p>{{transactionId}}</p>
             </div>
           </div>
           <div class="row w-75">
